+++ src/config.js
@@ -72,7 +72,7 @@ const config = {
     'convertOldData': {
         description: 'If you came from a version previous to <code>0.3.0</code>, you most probably have the old data in the atom folder. By default, it will always check on startup for this data and if the new does not exist, it will convert to the new data schema.',
         type: 'boolean',
-        default: true,
+        default: false,
         order: 10
     },
     'keepContext': {
@@ -99,11 +99,17 @@ const config = {
         default: '',
         order: 14
     },
+    'customTitleColor': {
+        description: 'Only allows for hexadecimal colors',
+        type: 'string',
+        default: '',
+        order: 15
+    },
     'onlyIcons': {
       description: 'Show the icons without the description.',
       type: 'boolean',
       default: false,
-      order: 15
+      order: 16
     }
 };