+++ src/list-item-component.js
@@ -76,8 +76,6 @@ function clickListener(evt) {
         atom.project.setPaths(model.projectPaths);
     }
 
-    // atom.devMode = model.projectDev
-
     _utility.setSelectedProjectView(this);
 
     if (selected) {